+++ src/components/GenerateSquares.js
@@ -1,0 +1,18 @@
+export const generateSquares = () => {
+  const squares = [];
+
+  let currentId = 0;
+
+  for (let row = 0; row < 3; row += 1) {
+    squares.push([]);
+    for (let col = 0; col < 3; col += 1) {
+      squares[row].push({
+        id: currentId,
+        value: '',
+      });
+      currentId += 1;
+    }
+  }
+
+  return squares;
+}
