+++ src/components/Square.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 
 import './Square.css'
 
-// Distruturing props
+// Distructuring props
 const Square = ({id, value, onClickCallback}) => {
   // For Wave 1 enable this 
   //  Component to alert a parent
